chore: move FCs to separate files as best practice from stencil
