+++ packages/amplify-ui-components/src/components/amplify-sign-in/amplify-forgot-password-hint.tsx
@@ -1,0 +1,8 @@
+import { FunctionalComponent as FC, h } from '@stencil/core';
+import { AmplifyForgotPasswordHintProps } from './amplify-sign-in-interface';
+
+export const AmplifyForgotPasswordHint: FC<AmplifyForgotPasswordHintProps> = ({ forgotPasswordText, resetPasswordText }) => (
+  <div>
+    {forgotPasswordText} <amplify-link>{resetPasswordText}</amplify-link>
+  </div>
+);
